requirements.txt
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.2
cycler==0.12.1
Django==4.2.7
et_xmlfile==2.0.0
fonttools==4.58.2
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.66.0
h11==0.14.0
idna==3.10
joblib==1.5.1
kiwisolver==1.4.8
matplotlib==3.10.3
mysqlclient==2.2.7
numpy==2.2.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.0
pyasn1-modules==0.4.1
pycparser==2.22
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
resend==0.10.2
sqlparse==0.5.3
statsmodels==0.14.4
threadpoolctl==3.6.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.3.0
